fix(privacy): move PrivacyExportSizeLimitExceededException to Exceptions/ subfolder
Architecture rule Exception_classes_should_reside_in_Exceptions_folder requires
exceptions to live under an Exceptions/ folder. The previous location at
DataExport/PrivacyExportSizeLimitExceededException.cs satisfied 'no public types
in Internal namespaces' but tripped this companion rule.

Final namespace: Granit.Privacy.BlobStorage.DataExport.Exceptions
Importers updated: CountingStream (throw site), ExportArchiveAssemblyHandler
(catch site), and the unit-test fixture.
