+++ tests/Granit.Privacy.BlobStorage.Tests/DataExport/Internal/CountingStreamTests.cs
@@ -1,4 +1,4 @@
-using Granit.Privacy.BlobStorage.DataExport;
+using Granit.Privacy.BlobStorage.DataExport.Exceptions;
 using Granit.Privacy.BlobStorage.DataExport.Internal;
 using Shouldly;
 using Xunit;
